+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Update `TransactionDataListItem` to receive `amountUsd` for transaction price, remove calculation
+
 ## [1.0.61] - 2024-12-29
 
 ### Added